Implement modular division of javascript functions
